Updated admin and user pages to be accessed only when logged in

+++ src/main/java/se/iths/webshop/ui/WebShopController.java
@@ -6,6 +6,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import se.iths.webshop.business.Customer;
+import se.iths.webshop.business.Employee;
 import se.iths.webshop.business.Person;
 import se.iths.webshop.business.WebShopService;
 
@@ -41,7 +43,8 @@ public class WebShopController {
         model.addAttribute("products", webShopService.getProducts());
         if (login == "Wrong password or email!") {
             return "login";
-        } else if (login == "admin") {
+        } else if (login == "Admin") {
+            model.addAttribute("login", webShopService.getUser().getName());
             return "admin";
         } else {
             return "shop";
@@ -50,48 +53,76 @@ public class WebShopController {
 
     @GetMapping("/shop")
     public String shopView(Model model) {
-        model.addAttribute("login", webShopService.getUser().getName());
-        model.addAttribute("products", webShopService.getProducts());
-        return "shop";
+        if (webShopService.getUser() instanceof Customer) {
+            model.addAttribute("login", webShopService.getUser().getName());
+            model.addAttribute("products", webShopService.getProducts());
+            return "shop";
+        }
+        model.addAttribute("login", "Please log in first");
+        return "login";
     }
 
     @GetMapping("/cart-view")
     public String cartView(Model model) {
-        model.addAttribute("login", webShopService.getUser().getName());
-        model.addAttribute("items", webShopService.getCart());
-        return "cart-view";
+        if (webShopService.getUser() instanceof Customer) {
+            model.addAttribute("login", webShopService.getUser().getName());
+            model.addAttribute("items", webShopService.getCart());
+            return "cart-view";
+        }
+        model.addAttribute("login", "Please log in first");
+        return "login";
     }
 
     @GetMapping("/logout")
     public String logoutUser(Model model) {
         String logout = webShopService.logoutUser();
-        model.addAttribute("loginResult", logout);
+        model.addAttribute("login", logout);
         return "login";
     }
 
     @GetMapping("admin-add")
     public String adminViewAdd(Model model) {
-        return "admin";
+        if (webShopService.getUser() instanceof Employee) {
+            model.addAttribute("login", "Admin " + webShopService.getUser().getName());
+            return "admin";
+        }
+        model.addAttribute("login", "Please log in first");
+        return "login";
     }
 
     @PostMapping("/add")
     public String addProduct(Model model, @RequestParam String name, @RequestParam String category, @RequestParam double price) {
-        String add = webShopService.addProduct(name, category, price);
-        model.addAttribute("addResult", add);
-        return "admin";
+        if (webShopService.getUser() instanceof Employee) {
+            String add = webShopService.addProduct(name, category, price);
+            model.addAttribute("addResult", add);
+            model.addAttribute("login", webShopService.getUser().getName());
+            return "admin";
+        }
+        model.addAttribute("login", "Please log in first");
+        return "login";
     }
 
     @PostMapping("/remove")
     public String removeProduct(Model model, @RequestParam long id) {
-        String remove = webShopService.removeProduct(id);
-        model.addAttribute("products", webShopService.getProducts());
-        model.addAttribute("removeResult", remove);
-        return "admin-view";
+        if (webShopService.getUser() instanceof Employee) {
+            String remove = webShopService.removeProduct(id);
+            model.addAttribute("login", webShopService.getUser().getName());
+            model.addAttribute("products", webShopService.getProducts());
+            model.addAttribute("remove", remove);
+            return "admin-view";
+        }
+        model.addAttribute("login", "Please log in first");
+        return "login";
     }
 
     @GetMapping("/admin-view")
     public String adminViewAll(Model model) {
-        model.addAttribute("products", webShopService.getProducts());
-        return "admin-view";
+        if (webShopService.getUser() instanceof Employee) {
+            model.addAttribute("products", webShopService.getProducts());
+            model.addAttribute("login", webShopService.getUser().getName());
+            return "admin-view";
+        }
+        model.addAttribute("login", "Please log in first");
+        return "login";
     }
 }
